priorities added
