+++ src/Batch/Program.cs
@@ -8,6 +8,7 @@ using AMTools.Shared.Core.Models;
 using AMTools.Shared.Core.Services.Interfaces;
 using AMTools.Shared.Core.Services.VirtualDesktops.Interfaces;
 using AMTools.Web.Core.ExtensionMethods;
+using AMTools.Web.Core.Services.Interfaces;
 using AMTools.Web.Data.Database;
 using AMTools.Web.Data.Database.Models;
 using McMaster.Extensions.CommandLineUtils;
