+++ src/Web/Startup.cs
@@ -57,6 +57,9 @@ namespace AMTools.Web
             services.InjectBackgroundServices();
 
             services.AddMemoryCache(action => { action.CompactionPercentage = .25; });
+
+            // TODO Prio 4: Frontend hinzufügen
+            // Fernsteuerungsseite mit Pfeilen <-- -->
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
